City tabs changes
